move get_vertexcolor tests from test_html_surface to test_matplotlib_backend
